:lipstick: add more changes and feats to profile pg
I'm realizing that the styles for profiles for the LOGGED IN user and
OTHER users are going to be drastically different. E.x. there won't be a "Follow" or
"Message" button on the logged in user's profile bc it wouldn't make
sense. So there's going to be varying components and a lot of dynamic rendering
going on.

I also setup some dummy data in the profile page just for the sake of
making the site interactive and planning out how I'm going to render the
data once I begin pulling from an actual API.